Added: ObjectsHierarchy

+++ Vector2D.cpp
@@ -7,7 +7,7 @@ Vector2D::Vector2D()
 	y = 0;
 }
 
-Vector2D::Vector2D(int x, int y)
+Vector2D::Vector2D(double x, double y)
 {
 	this->x = x;
 	this->y = y;
